KiraAI/requirements.txt
typer~=0.12.3
typing_extensions~=4.11.0
loguru~=0.7.2
python-dotenv~=1.0.1
pyftg~=2.1.1
pyftg-sound~=0.2
librosa~=0.10.2.post1
numpy~=1.26.4
tensorflow~=2.16.1
tqdm~=4.66.4